refactor: extracts `sorting-node` into its own file
