+++ typings/index.ts
@@ -1,11 +1,1 @@
-import type { TSESTree } from '@typescript-eslint/types'
-
-export interface SortingNode<Node extends TSESTree.Node = TSESTree.Node> {
-  hasMultipleImportDeclarations?: boolean
-  addSafetySemicolonWhenInline?: boolean
-  isEslintDisabled: boolean
-  group?: string
-  name: string
-  size: number
-  node: Node
-}
+export type { SortingNode } from './sorting-node'
