+++ typings/sorting-node.ts
@@ -1,0 +1,11 @@
+import type { TSESTree } from '@typescript-eslint/types'
+
+export interface SortingNode<Node extends TSESTree.Node = TSESTree.Node> {
+  hasMultipleImportDeclarations?: boolean
+  addSafetySemicolonWhenInline?: boolean
+  isEslintDisabled: boolean
+  group?: string
+  name: string
+  size: number
+  node: Node
+}
